% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr_report.R
\name{enr_report}
\alias{enr_report}
\title{Template for ENR report}
\usage{
enr_report(before_body)
}
\arguments{
\item{before_body}{html file you want to add. Specifically, the report uses information from _header.html and _tooltip_content.html.}
}
\description{
Template for ENR report
}
