% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_readability.R
\name{mean_readability}
\alias{mean_readability}
\title{Function that calculates readability index given link to the markdown file
Calculate the average of the 5 readability scores:
Flesch Kincaid, Gunning Fog Index, Coleman Liau, SMOG, Automated Readability Index}
\usage{
mean_readability(link)
}
\arguments{
\item{link}{a link of Rmd document}
}
\description{
Function that calculates readability index given link to the markdown file
Calculate the average of the 5 readability scores:
Flesch Kincaid, Gunning Fog Index, Coleman Liau, SMOG, Automated Readability Index
}
