% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publish file or folder on kvm1 server}
\usage{
publish(file = "", dest_folder = "")
}
\arguments{
\item{file}{name of a file in report folder that you want to send to server. Default value: '', then whole 'report' folder will be published.}

\item{dest_folder}{name of the folder in which files ought be saved. Default value: '', then name of folder where the project is located will be used.}
}
\description{
Save a file or a folder on the kvm1 server.
}
\examples{
\dontrun{
# save the whole 'report' catalogue on the server
upload_to_server() 

# save 'slide_deck.html' file located in 'report/slide_deck_template' folder to folder with a same name as project            
upload_to_server('slide_deck_template/slide_deck.html') 

# save 'slide_deck.html' located in 'report/slide_deck_template' to folder 'why_r'
upload_to_server(file = 'slide_deck_template/slide_deck.html', dest_folder = 'why_r')     
     }
}
