% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_index.R
\name{render_index}
\alias{render_index}
\title{Function which render index.Rmd-related files in project}
\usage{
render_index(index_file = "index", subdir = "reports")
}
\arguments{
\item{index_file}{Rmd index file name which we want to render.}

\item{subdir}{subdirectory name where we want to store our report. Default value: 'reports'}
}
\description{
Function which render index.Rmd-related files in project
}
\examples{
\dontrun{
render_all(index_file = "index", subdir = "reports")
}

}
