% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_create.R
\name{report_create}
\alias{report_create}
\title{Function which bulid template for ENR report}
\usage{
report_create(file, subdir = "reports")
}
\arguments{
\item{file}{Rmd file name which we want to create.}

\item{subdir}{subdirectory name where we want to store our report. Default value: 'reports'}
}
\description{
Function which bulid template for ENR report
}
\examples{
\dontrun{
report_create('001_initial_report')
     }
}
