% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reppear.R
\docType{package}
\name{reppear-package}
\alias{reppear}
\alias{reppear-package}
\title{reppear: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
}

}
\keyword{internal}
