% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reppear.R
\docType{package}
\name{reppear-package}
\alias{reppear}
\alias{reppear-package}
\title{reppear: A package which allows to build and maintain a project}
\description{
More about what it does (maybe more than one line)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pearsonplc/reppear}
  \item Report bugs at \url{https://github.com/pearsonplc/reppear/issues}
}

}
\keyword{internal}
